Made minor updates to the codebase
- Replaced usage of 'new' and 'delete' in state managers with implementation of unique pointers

- Added safety checks and error messages for asset managers that use map indexing

- Updated 'ApproxEq' function with absolute epsilon value

- Updated 'Animation' class to use uint32_t for animation timing

- Added 'internal' directory and updated absolute paths for the log and save files

- Moved all asset file paths to 'internal/assets.json'

- Added 'LoadJson' and 'OverWriteJson' helper functions for animation loading, asset loading, and the save system

+++ sources/Media/Animation.cpp
@@ -61,7 +61,7 @@ void Animation::NextFrame()
 
 void Animation::StepTime()
 {
-    uint8_t dt = (SDL_GetTicks() - m_currentTime);
+    uint32_t dt = (SDL_GetTicks() - m_currentTime);
 
     m_elapsedTime += dt;
 }
